requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
sentence-transformers
torch
pypdf
Pillow
